include/evmc/evmc.hpp: review type part 16

+++ include/evmc/evmc.hpp
@@ -313,7 +313,7 @@ constexpr address operator""_address(const char* s, size_t) noexcept
 
 #else
 /// Literal for evmc::address.
-constexpr address operator""_address(const char* s, unsigned int) noexcept
+constexpr address operator""_address(const char* s, unsigned long) noexcept
 {
     return parse<address>(s, "Z");
 }
